Add standalone /request-access page, mobile-optimized
Extracts form logic into shared AccessForm component used by both the
new standalone page and the existing Commercial section on /teams.
Header CTA and all "Get early access" links now point to /request-access.
