+++ apps/landing/src/App.tsx
@@ -3,6 +3,7 @@ import { Layout } from '@/components/Layout';
 import HomePage from '@/pages/HomePage';
 import TeamsPage from '@/pages/TeamsPage';
 import KnowledgePage from '@/pages/KnowledgePage';
+import RequestAccessPage from '@/pages/RequestAccessPage';
 
 export default function App() {
   return (
@@ -12,6 +13,7 @@ export default function App() {
           <Route path="/" element={<HomePage />} />
           <Route path="/teams" element={<TeamsPage />} />
           <Route path="/knowledge" element={<KnowledgePage />} />
+          <Route path="/request-access" element={<RequestAccessPage />} />
           <Route path="*" element={<HomePage />} />
         </Routes>
       </Layout>
